修改Bug,准备打包

+++ Assets/Scripe/SkillPakeger/skillAttack/skillBackpack/PackbackSkill400001.cs
@@ -3,7 +3,7 @@ public class PackbackSkill400001 : PackbackSkillBase
 {
     public override void removeSkill()
     {
-        mManager.carHurtPre -= value / 10000;
+        mManager.carHurtPre -= ((float)value / 10000);
     }
 
     public override void startSkill()
