+++ Assets/Scripe/SkillPakeger/skillAttack/skillBackpack/PackbackSkill400007.cs
@@ -3,11 +3,11 @@ public class PackbackSkill400007 : PackbackSkillBase
 {
     public override void removeSkill()
     {
-        mManager.cardCardHurtPre -= value / 10000;
+        mManager.cardCardHurtPre -= ((float)value / 10000);
     }
 
     public override void startSkill()
     {
-        mManager.cardCardHurtPre += value / 10000;
+        mManager.cardCardHurtPre += ((float)value / 10000);
     }
 }
